move logging utils to other class
